// src/app/models/project.model.ts
export class ProjectModel {
     id: number | undefined;
     ma: string | undefined;
     ten: string | undefined;
     tienDo: number | undefined;
     chiPhi: number | undefined;
     ngayBatDau: string | undefined;
     ngayKetThuc: string | undefined;
     ghiChu: string | undefined;
     trangThai: number | undefined;
}